moved calculate wedge cut to utils
